[GPU][CPU] Move RoPE to common part
