+++ src/common/transformations/include/ov_ops/rotary_positional_embeddings.hpp
@@ -1,4 +1,4 @@
-// Copyright (C) 2018-2024 Intel Corporation
+// Copyright (C) 2024 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 //
 
@@ -14,23 +14,37 @@ namespace internal {
 ///
 /// \brief Rotary Positional Embeddings operation
 /// Internal operation which may change in the future
-/// \ingroup ov_ops_cpp_api
-class TRANSFORMATIONS_API RPE : public ov::op::Op {
+class TRANSFORMATIONS_API RoPE : public Op {
 public:
-    OPENVINO_OP("RPE", "ie_internal_opset", op::Op);
+    OPENVINO_OP("RoPE", "ie_internal_opset", Op);
 
-    RPE() = default;
-    RPE(const Output<Node>& data, const Output<Node>& sin, const Output<Node>& cos, int64_t axis);
+    RoPE() = default;
 
-    void set_axis(int64_t axis);
-    int64_t get_axis() const;
+    struct Config {
+        size_t slice_start = 0;        // slice inner-most dimensions of input
+        size_t slice_stop = 0;
+        bool input_trans0213 = false;  // transpose input dim 1&2
+        bool is_interleaved = false;   // interleaved mode, implies trans0213 happens after RoPE
+        size_t rotary_ndims = 0;       // dimensions to be embedded (d in the description)
+        bool is_chatglm = false;       // chatglm is special which overrides other setting
+        bool is_qwen = false;          // Qwen is special which overrides other setting
+        size_t head_cnt = 0;
+        size_t head_size = 0;
+        int gather_position_arg_id =
+            0;  // arg id of position tensor, ==3 when gather from sin/cos inputs according to position is required
+    };
 
-    bool visit_attributes(AttributeVisitor& visitor) override;
+    RoPE(const OutputVector& args, const Config& config);
+
+    const Config& get_config() const;
+    void set_config(const Config& config);
+
+    bool visit_attributes(ov::AttributeVisitor& visitor) override;
     void validate_and_infer_types() override;
-    std::shared_ptr<Node> clone_with_new_inputs(const OutputVector& new_args) const override;
+    std::shared_ptr<Node> clone_with_new_inputs(const ov::OutputVector& new_args) const override;
 
 private:
-    int64_t m_axis{};
+    Config m_config{};
 };
 
 }  // namespace internal
